@@ -7,83 +7,87 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~15.1.0",
-    "@angular/cdk": "^15.1.4",
-    "@angular/common": "~15.1.0",
-    "@angular/compiler": "~15.1.0",
-    "@angular/core": "~15.1.0",
-    "@angular/forms": "~15.1.0",
-    "@angular/material": "^15.1.4",
-    "@angular/platform-browser": "~15.1.0",
-    "@angular/platform-browser-dynamic": "~15.1.0",
-    "@angular/router": "~15.1.0",
-    "@ngxs/logger-plugin": "^3.7.6",
-    "@ngxs/store": "^3.7.6",
-    "@nrwl/angular": "15.6.3",
-    "angular-material-css-vars": "^4.2.1",
-    "immer": "^9.0.19",
+    "@angular/animations": "15.2.6",
+    "@angular/cdk": "15.2.6",
+    "@angular/common": "15.2.6",
+    "@angular/compiler": "15.2.6",
+    "@angular/core": "15.2.6",
+    "@angular/forms": "15.2.6",
+    "@angular/material": "15.2.6",
+    "@angular/platform-browser": "15.2.6",
+    "@angular/platform-browser-dynamic": "15.2.6",
+    "@angular/router": "15.2.6",
+    "@ngxs/logger-plugin": "^3.8.0",
+    "@ngxs/store": "^3.8.0",
+    "@nrwl/angular": "15.9.2",
+    "angular-material-css-vars": "^4.2.3",
+    "immer": "^9.0.21",
     "js-yaml": "^4.1.0",
     "ng-inline-svg-2": "^15.0.1",
     "ngx-google-analytics": "^14.0.1",
-    "ngx-markdown": "^15.1.1",
+    "ngx-markdown": "^15.1.2",
     "normalize.css": "^8.0.1",
-    "papaparse": "^5.3.2",
-    "rxjs": "~7.5.0",
+    "papaparse": "^5.4.1",
+    "rxjs": "7.8.0",
     "tslib": "^2.3.0",
     "zod": "^3.21.4",
-    "zone.js": "~0.11.4"
+    "zone.js": "0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~15.1.0",
-    "@angular-eslint/eslint-plugin": "~15.0.0",
-    "@angular-eslint/eslint-plugin-template": "~15.0.0",
-    "@angular-eslint/template-parser": "~15.0.0",
-    "@angular/cli": "~15.1.0",
-    "@angular/compiler-cli": "~15.1.0",
-    "@angular/language-service": "~15.1.0",
+    "@angular-devkit/build-angular": "15.2.5",
+    "@angular-devkit/core": "15.2.5",
+    "@angular-devkit/schematics": "15.2.5",
+    "@angular-eslint/eslint-plugin": "^15.2.1",
+    "@angular-eslint/eslint-plugin-template": "^15.2.1",
+    "@angular-eslint/template-parser": "^15.2.1",
+    "@angular/cli": "~15.2.0",
+    "@angular/compiler-cli": "15.2.6",
+    "@angular/language-service": "15.2.6",
     "@compodoc/compodoc": "^1.1.19",
-    "@nrwl/cypress": "15.6.3",
-    "@nrwl/eslint-plugin-nx": "15.6.3",
-    "@nrwl/jest": "15.6.3",
-    "@nrwl/linter": "15.6.3",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/storybook": "^15.6.3",
-    "@nrwl/web": "^15.6.3",
-    "@nrwl/workspace": "15.6.3",
-    "@storybook/addon-essentials": "^6.5.15",
-    "@storybook/angular": "^6.5.15",
-    "@storybook/builder-webpack5": "^6.5.15",
-    "@storybook/core-common": "^6.5.15",
-    "@storybook/core-server": "^6.5.15",
-    "@storybook/manager-webpack5": "^6.5.15",
-    "@storybook/test-runner": "^0.7.2",
+    "@nrwl/cypress": "15.9.2",
+    "@nrwl/eslint-plugin-nx": "15.9.2",
+    "@nrwl/jest": "15.9.2",
+    "@nrwl/linter": "15.9.2",
+    "@nrwl/nx-cloud": "15.3.5",
+    "@nrwl/storybook": "15.9.2",
+    "@nrwl/web": "15.9.2",
+    "@nrwl/workspace": "15.9.2",
+    "@schematics/angular": "^15.2.5",
+    "@storybook/addon-essentials": "^7.0.2",
+    "@storybook/angular": "^7.0.2",
+    "@storybook/core-common": "^7.0.2",
+    "@storybook/core-server": "^7.0.2",
+    "@storybook/test-runner": "^0.10.0",
     "@twittwer/compodoc": "^1.9.0",
-    "@types/jest": "28.1.1",
+    "@types/jest": "^29.5.0",
     "@types/js-yaml": "^4.0.5",
-    "@types/node": "16.11.7",
+    "@types/node": "^18.15.11",
     "@types/papaparse": "^5.3.7",
-    "@typescript-eslint/eslint-plugin": "^5.36.1",
-    "@typescript-eslint/parser": "^5.36.1",
-    "cypress": "^12.2.0",
-    "eslint": "~8.15.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
+    "@typescript-eslint/eslint-plugin": "^5.58.0",
+    "@typescript-eslint/parser": "^5.58.0",
+    "cypress": "^12.9.0",
+    "eslint": "^8.38.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-cypress": "^2.13.2",
+    "eslint-plugin-storybook": "^0.6.11",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^8.0.3",
-    "jest": "28.1.1",
-    "jest-environment-jsdom": "28.1.1",
-    "jest-mock-extended": "^3.0.1",
-    "jest-preset-angular": "~12.2.3",
-    "lint-staged": "^13.1.1",
-    "ng-packagr": "~15.1.0",
-    "nx": "15.6.3",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "jest-mock-extended": "^3.0.4",
+    "jest-preset-angular": "^13.0.1",
+    "lint-staged": "^13.2.1",
+    "ng-packagr": "15.2.2",
+    "nx": "15.9.2",
     "postcss": "^8.4.5",
-    "postcss-import": "~14.1.0",
-    "postcss-preset-env": "~7.5.0",
+    "postcss-import": "^15.1.0",
+    "postcss-preset-env": "^8.3.1",
     "postcss-url": "~10.1.3",
-    "prettier": "^2.6.2",
-    "shallow-render": "^15.0.1",
-    "ts-jest": "28.0.5",
+    "prettier": "^2.8.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "shallow-render": "^15.0.2",
+    "ts-jest": "^29.1.0",
     "ts-node": "10.9.1",
     "typescript": "4.9.5",
     "webpack": "^5.64.0"
