@@ -1,6 +1,10 @@
-import { ResourceRegistryActions, ResourceRegistryState } from '@hra-ui/cdk/state';
-import { NgxsModule } from '@ngxs/store';
-import { moduleMetadata, Story, Meta } from '@storybook/angular';
+import {
+  LinkRegistryActions,
+  LinkRegistryState,
+  ResourceRegistryActions,
+  ResourceRegistryState,
+} from '@hra-ui/cdk/state';
+import { Meta, Story } from '@storybook/angular';
 import { HeaderBehaviorComponent } from './header-behavior.component';
 
 export default {
@@ -8,15 +12,13 @@ export default {
   component: HeaderBehaviorComponent,
   parameters: {
     state: {
-      states: [ResourceRegistryState],
-      actions: [new ResourceRegistryActions.LoadFromYaml('assets/resources/header.yml')],
+      states: [ResourceRegistryState, LinkRegistryState],
+      actions: [
+        new ResourceRegistryActions.LoadFromYaml('assets/resources/header.yml'),
+        new LinkRegistryActions.LoadFromYaml('assets/links/header.yml'),
+      ],
     },
   },
-  decorators: [
-    moduleMetadata({
-      imports: [NgxsModule.forRoot([ResourceRegistryState])],
-    }),
-  ],
 } as Meta<HeaderBehaviorComponent>;
 
 const Template: Story<HeaderBehaviorComponent> = (args: HeaderBehaviorComponent) => ({
