@@ -1,11 +1,16 @@
 import { Shallow } from 'shallow-render';
 
+import { LinkDirective } from '@hra-ui/cdk';
+import { selectQuerySnapshot } from '@hra-ui/cdk/injectors';
 import { HeaderComponent } from './header.component';
+
+jest.mock('@hra-ui/cdk/injectors');
+jest.mocked(selectQuerySnapshot).mockReturnValue(() => undefined);
 
 describe('HeaderComponent', () => {
   let shallow: Shallow<HeaderComponent>;
   beforeEach(() => {
-    shallow = new Shallow(HeaderComponent);
+    shallow = new Shallow(HeaderComponent).dontMock(LinkDirective);
   });
 
   it('renders images correctly', async () => {
